rename handler dir to service
