+++ service/result/handler.go
