+++ service/result/result.go
