+++ service/service_handler.go
@@ -1,4 +1,4 @@
-package handler
+package service
 
 import (
 	"fmt"
@@ -7,8 +7,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	cosmosrpc "github.com/cosmos/gogoproto/grpc"
 
-	result "github.com/0xPellNetwork/pellapp-sdk/handler/result"
-	"github.com/0xPellNetwork/pellapp-sdk/handler/tx"
+	"github.com/0xPellNetwork/pellapp-sdk/service/result"
+	"github.com/0xPellNetwork/pellapp-sdk/service/tx"
 	sdktypes "github.com/0xPellNetwork/pellapp-sdk/types"
 )
